inline comment boxes + physical LayoutRow model (C1(a)/(b) landed)
Replace the single-line `● you` comment row with a bordered multi-row box, and
land the deferred C1 physical-layout refactor the box needs.

Refactor (C1(a)/(b)):
- Vec<LayoutRow { target, subrow, side, hit }> is the physical layout backing the
  C1 cursor seam; review_target_at/review_index_of/review_cursor_span read it, with
  the cache keyed on (pane_width, diff_mode) and cleared on every diff/comment
  change. A code line = 1 row; a comment = title + wrapped body + bottom, all one
  RowTarget::Comment(id). review_cursor_highlight -> Option<Range<usize>> (full-box
  highlight). Scroll/content metrics migrated u16 -> usize.

Box render (§3.4, both modes):
- rounded theme.comment border; title `● <you|agent> — <file> R<line>` + a
  right-aligned `[x]` that stays visible when the title truncates; word-wrapped
  body via char_width. Unified = full-width; side-by-side = box in the anchor's
  side column with equal-height blanks in the other. Stale worktree notes render
  dim. x_rects/comment_close_rect recorded for C8's click.
- shared TestBackend colour-assert helper in tests/common; tests/comment_box_test.rs
  (10 cases) + updated single-line assertions across the comment suites to the box.

Review: the full-diff codex correctness pass hit its 12-minute cap (cancelled). In
its place: a targeted self-review of the highest-risk surface — the (width,mode)
layout cache — confirmed it is invalidated on all seven comment-mutation paths and
every diff/file change, so a stale layout can't misdirect a click/delete. Defensive
narrow-terminal edge-testing (real --dump-frame at widths 24..2) then FOUND + FIXED
a subtract-with-overflow panic: box_body_spans guarded `box_w < 2` but subtracted 4;
now guards `box_w < 4` (degrades to a bare border column). +regression test.

Verified visually via a real --dump-frame box artifact (SBS side-column placement,
wrapping, [x]) in the plan's artifact folder.

+++ tests/comment_input_test.rs
@@ -511,12 +511,12 @@ fn c_on_an_offscreen_cursor_reveals_without_opening() {
     for _ in 0..12 {
         app.on_mouse(mouse(diff.x + 2, diff.y + 1, MouseEventKind::ScrollDown));
     }
-    assert!(app.diff_scroll as usize > cursor, "the cursor is offscreen");
+    assert!(app.diff_scroll > cursor, "the cursor is offscreen");
 
     app.on_key(key('c'));
     assert!(!modal_open(&app), "the first c only reveals — no editor");
     assert!(
-        (app.diff_scroll as usize) <= cursor,
+        app.diff_scroll <= cursor,
         "the cursor's row was scrolled back into view"
     );
